fix[cert-manager-setup]: deploy Issuer CRs in the release namespace
Prior to this fix, the namespace is set to an empty string. When the namespace
is an empty string, helm v2 deploys the Issuer CR in the release namespace, but
helm v3 deploys it in namespace where the the kubeaddons-controller runs.
That's a problem, because cert-manager looks for the Issuers in its own
namespace.